+++ source/Energinet.DataHub.MeteringPoints.Tests/Domain/MasterDataHandling/UpdateTests.cs
@@ -14,15 +14,12 @@
 
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling;
 using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Components;
 using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Components.Addresses;
 using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Components.MeteringDetails;
 using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Errors;
 using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Rules;
-using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints;
-using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 using Xunit;
 using Xunit.Categories;
 
@@ -290,231 +287,6 @@ namespace Energinet.DataHub.MeteringPoints.Tests.Domain.MasterDataHandling
         private MasterDataUpdater UpdateBuilder(MasterData current, IEnumerable<MasterDataField> fieldConfiguration = null)
         {
             return new MasterDataUpdater(fieldConfiguration ?? new List<MasterDataField>(), current);
-        }
-    }
-
-    public class MasterDataUpdater : MasterDataBuilderBase, IMasterDataBuilder
-    {
-        private readonly List<ValidationError> _validationErrors = new();
-
-        public MasterDataUpdater(IEnumerable<MasterDataField> fields, MasterData currentMasterData)
-            :base(fields)
-        {
-            SetValue(nameof(MasterData.ProductType), currentMasterData.ProductType);
-            SetValue(nameof(MasterData.NetSettlementGroup), currentMasterData.NetSettlementGroup);
-            SetValue(nameof(MasterData.ConnectionType), currentMasterData.ConnectionType);
-            SetValue(nameof(MasterData.MeteringConfiguration), currentMasterData.MeteringConfiguration);
-        }
-
-        public BusinessRulesValidationResult Validate()
-        {
-            _validationErrors.Clear();
-            _validationErrors.AddRange(AllValueValidationErrors());
-
-            return new BusinessRulesValidationResult(_validationErrors);
-        }
-
-        private void RemoveValueIfNotApplicable<T>(string valueName, Func<bool> rule)
-        {
-            if (rule())
-            {
-                SetValue<T>(valueName, default(T));
-            }
-        }
-
-        public MasterData Build()
-        {
-            RemoveValueIfNotApplicable<ScheduledMeterReadingDate>(nameof(MasterData.ScheduledMeterReadingDate),
-                () => GetValue<NetSettlementGroup>(nameof(MasterData.NetSettlementGroup)) != NetSettlementGroup.Six);
-            RemoveValueIfNotApplicable<ConnectionType>(nameof(MasterData.ConnectionType),
-                () => GetValue<NetSettlementGroup>(nameof(MasterData.NetSettlementGroup)) == NetSettlementGroup.Zero);
-
-            return new MasterData(
-                productType: GetValue<ProductType>(nameof(MasterData.ProductType)),
-                unitType: GetValue<MeasurementUnitType>(nameof(MasterData.UnitType)),
-                assetType: GetValue<AssetType>(nameof(MasterData.AssetType)),
-                readingOccurrence: GetValue<ReadingOccurrence>(nameof(MasterData.ReadingOccurrence)),
-                powerLimit: GetValue<PowerLimit>(nameof(MasterData.PowerLimit)),
-                powerPlantGsrnNumber: GetValue<GsrnNumber>(nameof(MasterData.PowerPlantGsrnNumber)),
-                effectiveDate: GetValue<EffectiveDate>(nameof(MasterData.EffectiveDate)),
-                capacity: GetValue<Capacity>(nameof(MasterData.Capacity)),
-                address: GetValue<Address>(nameof(MasterData.Address)),
-                meteringConfiguration: GetValue<MeteringConfiguration>(nameof(MasterData.MeteringConfiguration)),
-                settlementMethod: GetValue<SettlementMethod>(nameof(MasterData.SettlementMethod)),
-                scheduledMeterReadingDate: GetValue<ScheduledMeterReadingDate>(nameof(MasterData.ScheduledMeterReadingDate)),
-                connectionType: GetValue<ConnectionType>(nameof(MasterData.ConnectionType)),
-                disconnectionType: GetValue<DisconnectionType>(nameof(MasterData.DisconnectionType)),
-                netSettlementGroup: GetValue<NetSettlementGroup>(nameof(MasterData.NetSettlementGroup)),
-                productionObligation: GetValue<bool?>(nameof(MasterData.ProductionObligation)));
-        }
-
-        public IMasterDataBuilder WithNetSettlementGroup(string netSettlementGroup)
-        {
-            SetValueIfValid(
-                nameof(MasterData.NetSettlementGroup),
-                () =>
-                {
-                    return EnumerationType.GetAll<NetSettlementGroup>()
-                        .Select(item => item.Name)
-                        .Contains(netSettlementGroup) == false ? BusinessRulesValidationResult.Failure(new InvalidSettlementGroupValue(netSettlementGroup)) : BusinessRulesValidationResult.Valid();
-                },
-                () => EnumerationType.FromName<NetSettlementGroup>(netSettlementGroup));
-            return this;
-        }
-
-        public IMasterDataBuilder WithMeteringConfiguration(string? method, string? meterNumber)
-        {
-            var currentMeterConfiguration = GetValue<MeteringConfiguration>(nameof(MasterData.MeteringConfiguration));
-            SetValueIfValid(
-                nameof(MasterData.MeteringConfiguration),
-                () =>
-                {
-                    if (method is not null)
-                    {
-                        if (EnumerationType.GetAll<MeteringMethod>()
-                            .Select(item => item.Name)
-                            .Contains(method) == false)
-                        {
-                            return BusinessRulesValidationResult.Failure(new InvalidMeteringMethodValue(method));
-                        }
-                    }
-
-                    if (string.IsNullOrWhiteSpace(meterNumber) == false)
-                    {
-                        var meterValidationResult = MeterId.CheckRules(meterNumber);
-                        if (meterValidationResult.Success == false)
-                        {
-                            return BusinessRulesValidationResult.Failure(meterValidationResult.Errors.ToArray());
-                        }
-                    }
-
-                    var meteringMethod = method is null
-                        ? currentMeterConfiguration.Method
-                        : EnumerationType.FromName<MeteringMethod>(method);
-                    BusinessRulesValidationResult validationResult;
-                    if (meteringMethod == MeteringMethod.Physical)
-                    {
-                        validationResult = MeteringConfiguration.CheckRules(meteringMethod, string.IsNullOrEmpty(meterNumber) ? MeterId.Empty() : MeterId.Create(meterNumber));
-                    }
-                    else
-                    {
-                        validationResult = MeteringConfiguration.CheckRules(meteringMethod, MeterId.Empty());
-                    }
-
-                    if (validationResult.Success == false)
-                    {
-                        return BusinessRulesValidationResult.Failure(validationResult.Errors.ToArray());
-                    }
-
-                    return BusinessRulesValidationResult.Valid();
-                },
-                () =>
-                {
-                    var meteringMethod = string.IsNullOrEmpty(method) ? currentMeterConfiguration.Method : EnumerationType.FromName<MeteringMethod>(method);
-                    var meterId = meteringMethod == MeteringMethod.Physical ? string.IsNullOrEmpty(meterNumber)
-                        ? currentMeterConfiguration.Meter : MeterId.Create(meterNumber) : MeterId.Empty();
-                    return MeteringConfiguration.Create(meteringMethod, meterId);
-                });
-            return this;
-        }
-
-        public IMasterDataBuilder WithAddress(string? streetName = null, string? streetCode = null, string? buildingNumber = null,
-            string? city = null, string? citySubDivision = null, string? postCode = null, CountryCode? countryCode = null,
-            string? floor = null, string? room = null, int? municipalityCode = null, bool? isActual = null,
-            Guid? geoInfoReference = null, string? locationDescription = null)
-        {
-            throw new NotImplementedException();
-        }
-
-        public IMasterDataBuilder WithMeasurementUnitType(string? measurementUnitType)
-        {
-            throw new NotImplementedException();
-        }
-
-        public IMasterDataBuilder WithPowerPlant(string? gsrnNumber)
-        {
-            throw new NotImplementedException();
-        }
-
-        public IMasterDataBuilder WithReadingPeriodicity(string? readingPeriodicity)
-        {
-            throw new NotImplementedException();
-        }
-
-        public IMasterDataBuilder WithPowerLimit(int kwh, int ampere)
-        {
-            throw new NotImplementedException();
-        }
-
-        public IMasterDataBuilder WithSettlementMethod(string? settlementMethod)
-        {
-            throw new NotImplementedException();
-        }
-
-        public IMasterDataBuilder WithDisconnectionType(string? disconnectionType)
-        {
-            throw new NotImplementedException();
-        }
-
-        public IMasterDataBuilder WithAssetType(string? assetType)
-        {
-            throw new NotImplementedException();
-        }
-
-        public IMasterDataBuilder WithScheduledMeterReadingDate(string? scheduledMeterReadingDate)
-        {
-            throw new NotImplementedException();
-        }
-
-        public IMasterDataBuilder WithCapacity(double? capacity)
-        {
-            throw new NotImplementedException();
-        }
-
-        public IMasterDataBuilder EffectiveOn(string? effectiveDate)
-        {
-            throw new NotImplementedException();
-        }
-
-        public IMasterDataBuilder WithProductType(string? productType)
-        {
-            if (productType is null)
-            {
-                return this;
-            }
-
-            SetValueIfValid(nameof(MasterData.ProductType),
-                BusinessRulesValidationResult.Valid,
-                () => EnumerationType.FromName<ProductType>(productType));
-
-            return this;
-        }
-
-        public IMasterDataBuilder WithConnectionType(string? connectionType)
-        {
-            if (connectionType is null) return this;
-            if (connectionType == string.Empty) SetValue<ConnectionType>(nameof(MasterData.ConnectionType), null);
-            if (connectionType != string.Empty) SetValueIfValid(nameof(MasterData.ConnectionType),
-                () =>
-                {
-                    return EnumerationType.GetAll<ConnectionType>()
-                        .Select(item => item.Name)
-                        .Contains(connectionType) == false ? BusinessRulesValidationResult.Failure(new InvalidConnectionTypeValue(connectionType)) : BusinessRulesValidationResult.Valid();
-                },
-                () => EnumerationType.FromName<ConnectionType>(connectionType));
-            return this;
-
-        }
-
-        public IMasterDataBuilder WithProductionObligation(bool? productionObligation)
-        {
-            throw new NotImplementedException();
-        }
-
-        private void SetValueIfValid<T>(string valueName, Func<BusinessRulesValidationResult> validator, Func<T> creator)
-        {
-            var valueItem = GetMasterValueItem<T>(valueName);
-            valueItem.SetValue(validator, creator);
         }
     }
 }
